cdk/requirements.txt
attrs==19.3.0
autopep8==1.5.7
aws-cdk.assets==1.42.0
aws-cdk.aws-apigateway==1.42.0
aws-cdk.aws-applicationautoscaling==1.42.0
aws-cdk.aws-autoscaling==1.42.0
aws-cdk.aws-autoscaling-common==1.42.0
aws-cdk.aws-autoscaling-hooktargets==1.42.0
aws-cdk.aws-batch==1.42.0
aws-cdk.aws-certificatemanager==1.42.0
aws-cdk.aws-cloudformation==1.42.0
aws-cdk.aws-cloudfront==1.42.0
aws-cdk.aws-cloudwatch==1.42.0
aws-cdk.aws-codebuild==1.42.0
aws-cdk.aws-codecommit==1.42.0
aws-cdk.aws-codepipeline==1.42.0
aws-cdk.aws-cognito==1.42.0
aws-cdk.aws-ec2==1.42.0
aws-cdk.aws-ecr==1.42.0
aws-cdk.aws-ecr-assets==1.42.0
aws-cdk.aws-ecs==1.42.0
aws-cdk.aws-ecs-patterns==1.42.0
aws-cdk.aws-elasticloadbalancing==1.42.0
aws-cdk.aws-elasticloadbalancingv2==1.42.0
aws-cdk.aws-events==1.42.0
aws-cdk.aws-events-targets==1.42.0
aws-cdk.aws-iam==1.42.0
aws-cdk.aws-kms==1.42.0
aws-cdk.aws-lambda==1.42.0
aws-cdk.aws-logs==1.42.0
aws-cdk.aws-route53==1.42.0
aws-cdk.aws-route53-targets==1.42.0
aws-cdk.aws-s3==1.42.0
aws-cdk.aws-s3-assets==1.42.0
aws-cdk.aws-sam==1.42.0
aws-cdk.aws-secretsmanager==1.42.0
aws-cdk.aws-servicediscovery==1.42.0
aws-cdk.aws-sns==1.42.0
aws-cdk.aws-sns-subscriptions==1.42.0
aws-cdk.aws-sqs==1.42.0
aws-cdk.aws-ssm==1.42.0
aws-cdk.aws-stepfunctions==1.42.0
aws-cdk.cdk-assets-schema==1.42.0
aws-cdk.cloud-assembly-schema==1.42.0
aws-cdk.core==1.42.0
aws-cdk.custom-resources==1.42.0
aws-cdk.cx-api==1.42.0
aws-cdk.region-info==1.42.0
cattrs==1.0.0
click==7.1.2
constructs==3.0.3
flake8==3.9.2
importlib-metadata==4.0.1
importlib-resources==5.1.2
jsii==1.5.0
mccabe==0.6.1
packaging==20.9
pep517==0.10.0
pip-review==1.1.0
pip-tools==6.1.0
publication==0.0.3
pycodestyle==2.7.0
pyflakes==2.3.1
pyparsing==2.4.7
python-dateutil==2.8.1
six==1.15.0
toml==0.10.2
typing-extensions==3.7.4.3
zipp==3.4.1
